Fix of finding out valid builds. Added retry button on mail tasks
[git-p4: depot-paths = "//depot/": change = 261]

+++ DbApp/site/views/entry/tmpl/mailqueue.php
@@ -5,16 +5,26 @@ defined('_JEXEC') or die('Restricted access'); ?>
   $menu  = $menus->getActive();
   $itemid = $menu->id;
   $searchid = JRequest::getVar('searchid', -1) ;
+  $action = JRequest::getVar('action', "");
   $mailtasks = $this->mailtasks;
 
   if ($searchid >= 0) {
-    $db =& JFactory::getDBO();
-    $query = " DELETE FROM #__aaafqmailqueue WHERE id=" . $db->Quote($searchid);
-    $db->setQuery($query);
-    if (! $db->query()) {
-      JError::raiseWarning('Message', JText::_('Could not remove task!'));
-	  $searchid = -1;
-    }
+      $db =& JFactory::getDBO();
+      if ($action == 'cancel') {
+          $query = " DELETE FROM #__aaafqmailqueue WHERE id=" . $db->Quote($searchid);
+          $db->setQuery($query);
+          if (! $db->query()) {
+              JError::raiseWarning('Message', JText::_('Could not remove task!'));
+	      $searchid = -1;
+          }
+      } else if ($action == 'retry') {
+          $query = ' UPDATE #__aaafqmailqueue SET status="inqueue" WHERE id=' . $db->Quote($searchid);
+          $db->setQuery($query);
+          if (! $db->query()) {
+              JError::raiseWarning('Message', JText::_('Could not reactivate task!'));
+	      $searchid = -1;
+          }
+      }
   }
 ?>
 <script type="text/javascript">
@@ -27,13 +37,17 @@ function confirmRemove() {
     <table><tr><th>Run&nbsp;</th><th>Lane&nbsp;</th><th>Recepient</th><th>Status</th><th></th></tr>
 <?php
   foreach ($mailtasks as $task) {
-    if ($task->id == $searchid) continue;
-    $cancellink = "";
-	if ($task->status == "inqueue")
-        $cancellink = "<a href=\"index.php?option=com_dbapp&view=entry&layout=mailqueue&controller=entry&searchid=" 
-                      . $task->id . "&Itemid=" . $itemid . "\" onclick=\"return confirmRemove();\">Remove</a>";
+    if ($task->id == $searchid && $action == 'cancel') continue;
+    if ($task->id == $searchid && $action == 'retry') $task->status = 'inqueue';
+    $actionlink = "";
+    if ($task->status == "inqueue")
+        $actionlink = "<a href=\"index.php?option=com_dbapp&view=entry&layout=mailqueue&controller=entry&searchid=" 
+                      . $task->id . "&action=cancel&Itemid=" . $itemid . "\" onclick=\"return confirmRemove();\">Remove</a>";
+    if ($task->status == "failed" || $task->status == "filemissing")
+        $actionlink = "<a href=\"index.php?option=com_dbapp&view=entry&layout=mailqueue&controller=entry&searchid="
+                      . $task->id . "&action=retry&Itemid=$itemid\">Retry</a>";
     echo "<tr><td>" . $task->runno . "</td><td>" . $task->laneno . "</td><td>" . $task->email 
-	     . "&nbsp;</td><td>" . $task->status . "&nbsp;</td><td>" . $cancellink . "&nbsp;</td></tr>\n";
+	     . "&nbsp;</td><td>" . $task->status . "&nbsp;</td><td>" . $actionlink . "&nbsp;</td></tr>\n";
   }
   echo "</table></fieldset></div>";
   echo "<br />\n";
